+++ CHANGES.rst
@@ -60,6 +60,8 @@ Specviz2d
 Bug Fixes
 ---------
 
+- Fixes in model fitting plugin to work better with nonfinite data. [#3792]
+
 Cubeviz
 ^^^^^^^
 
@@ -85,6 +87,7 @@ Other Changes and Additions
 - Linking is now generalized to act based on physical type. [#3698]
 
 - Improve memory usage when loading large cubes in cubeviz. [#3788]
+
 
 4.3.2 (2025-09-15)
 ==================